Remove custom echo_service and use the pigweed version Example cleanup
