--- examples/common/pigweed/protos/echo_service.options
@@ -1,2 +1,0 @@
-
-chip.rpc.EchoMessage.msg max_size:64
