--- examples/common/pigweed/protos/echo_service.proto
@@ -1,13 +1,0 @@
-syntax = "proto3";
-
-import 'pw_protobuf_protos/common.proto';
-
-package chip.rpc;
-
-message EchoMessage {
-  string msg = 1;
-}
-
-service Echo {
-  rpc Send(EchoMessage) returns (EchoMessage){}
-}
