--- examples/pigweed-app/mbed/main/include/Rpc.h
@@ -1,33 +1,0 @@
-// Copyright 2020 The Pigweed Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License"); you may not
-// use this file except in compliance with the License. You may obtain a copy of
-// the License at
-//
-//     https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
-// WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
-// License for the specific language governing permissions and limitations under
-// the License.
-#pragma once
-
-#include "echo_service/echo_service.rpc.pb.h"
-#include <cstring>
-
-namespace pw::rpc {
-
-} // namespace pw::rpc
-
-#pragma once
-
-#include "rtos/Thread.h"
-
-namespace chip {
-namespace rpc {
-
-void RunRpcService();
-
-} // namespace rpc
-} // namespace chip
